Move account routes into blueprint
and update templates,
and remove unused imports

Routes:
* /login
* /logout
* /register
* /profile
