+++ nyaa/templates/layout.html
@@ -128,13 +128,13 @@
 									</a>
 								</li>
 								<li>
-									<a href="{{ url_for('profile') }}">
+									<a href="{{ url_for('account.profile') }}">
 										<i class="fa fa-gear fa-fw"></i>
 										Profile
 									</a>
 								</li>
 								<li>
-									<a href="{{ url_for('logout') }}">
+									<a href="{{ url_for('account.logout') }}">
 										<i class="fa fa-times fa-fw"></i>
 										Logout
 									</a>
@@ -154,13 +154,13 @@
 							</a>
 							<ul class="dropdown-menu">
 								<li>
-									<a href="{{ url_for('login') }}">
+									<a href="{{ url_for('account.login') }}">
 										<i class="fa fa-sign-in fa-fw"></i>
 										Login
 									</a>
 								</li>
 								<li>
-									<a href="{{ url_for('register') }}">
+									<a href="{{ url_for('account.register') }}">
 										<i class="fa fa-pencil fa-fw"></i>
 										Register
 									</a>
