+++ nyaa/views/__init__.py
@@ -1,5 +1,7 @@
 from nyaa.views import (
+    account,
     site,
 )
 
+account_bp = account.bp
 site_bp = site.bp
